Plans Service: Phases and Milestones
(1) Create relationships between Phases and Milestones
(2) Create PlansModule
