+++ src/app.module.ts
@@ -1,14 +1,12 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
 
-import { MilestonesModule } from 'src/milestones/milestones.module';
-import { PhasesModule } from 'src/phases/phases.module';
+import { PlansModule } from 'src/plans/plans.module';
 
 @Module({
   imports: [
     MongooseModule.forRoot(process.env.MONGODB_URI as string, {}),
-    MilestonesModule,
-    PhasesModule,
+    PlansModule,
   ],
 })
 export class AppModule {}
