--- src/milestones/api/rest/dto/milestone.dto.ts
@@ -1,9 +1,0 @@
-export class CreateMilestoneDto {
-  readonly title: string;
-  readonly description: string;
-}
-
-export class UpdateMilestoneDto {
-  readonly title?: string;
-  readonly description?: string;
-}
