--- src/milestones/domain/entities/milestone.entity.ts
@@ -1,9 +1,0 @@
-export class Milestone {
-  constructor(
-    private readonly id: string,
-    private readonly title: string,
-    private readonly description: string,
-    private readonly createdAt?: string | boolean,
-    private readonly updatedAt?: string | boolean,
-  ) {}
-}
