+++ src/plans/milestones/api/rest/dto/milestone.dto.ts
@@ -1,0 +1,11 @@
+export class CreateMilestoneDto {
+  readonly title: string;
+  readonly description: string;
+  readonly phaseId: string;
+}
+
+export class UpdateMilestoneDto {
+  readonly title?: string;
+  readonly description?: string;
+  readonly phaseId?: string;
+}
