+++ src/plans/milestones/domain/entities/milestone.entity.ts
@@ -1,0 +1,22 @@
+export class PhaseId {
+  constructor(readonly id: string) {}
+}
+
+export class Milestone {
+  constructor(
+    readonly id: string,
+    readonly title: string,
+    readonly description: string,
+    readonly phaseId: PhaseId,
+    readonly createdAt?: string | boolean,
+    readonly updatedAt?: string | boolean,
+  ) {
+    if (!id) {
+      throw new Error('Milestone id is required');
+    }
+
+    if (!phaseId) {
+      throw new Error('Milestone phaseId is required');
+    }
+  }
+}
